Refactor import paths for LLM MCP clients to align with new structure. Update imports in various modules to reflect the move to a base client directory. Ensure compatibility across tests and utility files.
