--- src/datarobot_genai/drmcp/test_utils/clients/__init__.py
@@ -1,31 +1,0 @@
-# Copyright 2026 DataRobot, Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-"""LLM MCP Client implementations for various providers."""
-
-from .anthropic import AnthropicMCPClient
-from .base import BaseLLMMCPClient
-from .base import LLMResponse
-from .base import ToolCall
-from .dr_gateway import DRLLMGatewayMCPClient
-from .openai import OpenAILLMMCPClient
-
-__all__ = [
-    "AnthropicMCPClient",
-    "BaseLLMMCPClient",
-    "DRLLMGatewayMCPClient",
-    "LLMResponse",
-    "OpenAILLMMCPClient",
-    "ToolCall",
-]
