+++ src/datarobot_genai/drmcp/test_utils/utils.py
@@ -19,7 +19,7 @@ from typing import TYPE_CHECKING
 from typing import Any
 
 if TYPE_CHECKING:
-    from .clients import LLMResponse
+    from .clients.base import LLMResponse
 
 from dotenv import load_dotenv
